Further reduce allocations in adler32.Write
So I missed that our usage pattern, in one case, is a Reset(), Write(),
Reset(), Write() loop, which still causes an allocation for the window
on every iteration. This adds an ugly, nested if that fixes that. It's
not super beautiful, but it's a huge win in terms of allocations for
that pattern...